# docs/swagger.yaml
basePath: /api
definitions:
  api.APIErrorResponse:
    properties:
      diagnosisCode:
        description: diagnosis code of the error
        type: integer
      diagnosisMessage:
        description: diagnosis message of the error
        type: string
    type: object
  imagecontroller.SetPreviewImageRequestBody:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Swagger documentation to test the myInventory API
  title: myInventory API
  version: "1.0"
paths:
...
  /v1/items/{itemID}/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload one ore more images to the server and get their internal
        ids
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      - collectionFormat: csv
        description: images of the item
        in: formData
        items:
          type: file
        name: images
        required: true
        type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Upload images
      tags:
      - images
  /v1/items/{itemID}/images/{imageID}:
    delete:
      consumes:
      - application/json
      description: Delete an image from the item
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      - description: image id
        in: path
        name: imageID
        required: true
        type: string
      responses:
        "200":
          description: Image
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Delete an image
      tags:
      - images
    get:
      consumes:
      - application/json
      description: Get an image from the item
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      - description: image id
        in: path
        name: imageID
        required: true
        type: string
      responses:
        "200":
          description: Image
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Get an image
      tags:
      - images
  /v1/items/{itemID}/images/preview:
    put:
      consumes:
      - application/json
      description: Set preview image for an item
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      - description: Set Preview Image Name
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/imagecontroller.SetPreviewImageRequestBody'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Set preview image
      tags:
      - images
  /v1/items/{itemID}/invoice:
    delete:
      consumes:
      - application/json
      description: Delete the invoice from the item and from the storage
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Delete invoice
      tags:
      - invoice
    get:
      consumes:
      - application/json
      description: Get the invoice from the item
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "200":
          description: Invoice
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Get invoice
      tags:
      - invoice
    post:
      consumes:
      - multipart/form-data
      description: Upload invoice to the item
      parameters:
      - description: item id
        in: path
        name: itemID
        required: true
        type: string
      - description: invoice of the item
        in: formData
        name: invoice
        required: true
        type: file
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorResponse'
      security:
      - JWT: []
      summary: Upload invoice
      tags:
      - invoice
...
securityDefinitions:
  JWT:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"